feat: 요금에 따른 model 선택 정책 도입

+++ gss-domain/src/main/java/com/devoops/domain/entity/analysis/OpenAiModel.java
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 public enum OpenAiModel {
 
     GPT_5(0, 7500, "gpt-5", 0.00000125, 0.00001),
-    GPT_5_MINI(7501, 10000, "gpt-5-mini",0.00000025,0.000002),
+    GPT_5_MINI(7501, 10000, "gpt-5-mini", 0.00000025, 0.000002),
     GPT_5_NANO(10001, 15000, "gpt-5-nano", 0.00000005, 0.0000004),
     ;
 
@@ -20,9 +20,10 @@ public enum OpenAiModel {
     private final double inputTokenCharge; //달러
     private final double outputTokenCharge; //달러
 
-    public static OpenAiModel getModelByUsage(int currentUsageWon) {
+    public static OpenAiModel getModelByUsage(double currentUsageWon) {
         return Stream.of(values())
-                .filter(model -> model.moneyUnderCriteria<=currentUsageWon && model.moneyUpperCriteria>=currentUsageWon)
+                .filter(model -> model.moneyUnderCriteria <= currentUsageWon
+                        && model.moneyUpperCriteria >= currentUsageWon)
                 .findAny()
                 .orElse(GPT_5_NANO);
     }
